reorder the initial classes, so that they are back to the old ordering
So:   $classes, local/, local/$classes    (assuming use_local)

This lets one set site defaults in local/{preseed,subclass} and then
override them in local/$cls/preseed

Hopefully the case where append_classes is called with a parameter still
works, since we don't do the reordering then.  The motential for breakage
would seem to be that subclasses will be filtered at that stage if they
were already pulled in by the old_classes -- I suspect that if that's
a problem it is due to misuse of the subclasses, and would be possible
to work around by cloning the subclass in question.

locals_last: use sed to reorder local after non-local classes